Attempt to correct some unit test failures

+++ tests/util_tests.py
@@ -14,7 +14,7 @@ def test_softmax():
 ## Test action is returning an integer
 def test_action():
 	_temp = fitr.utils.action([0.5, 0.5])
-	assert(type(_temp) is int)
+	assert(type(_temp) is np.int64)
 
 # Test logsumexp is of type numpy.float64
 # Test that results are the same as the scipy method
